backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]>=0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
Pillow==10.3.0
google-genai==0.7.0
google-generativeai>=0.8.0
rapidfuzz==3.9.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
httpx>=0.25,<0.29
pydantic==2.7.1
email-validator>=2.0.0
pytest>=8.0.0
python-dotenv==1.0.1
pdf2image==1.17.0
spacy==3.7.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
medspacy==1.3.1
supabase>=2.10,<3
instructor==1.3.2
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
python-docx==1.1.2
PyMuPDF==1.24.10
reportlab==4.2.5
